fix(subq/in): split stream_subquery_in_vtable2 into tow files
